requirements.txt
SQLAlchemy==2.0.6
httpx==0.23.3
fastapi==0.95.1
pytest==7.2.2
asyncpg==0.27.0
greenlet==2.0.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
uvicorn==0.21.1
alembic==1.10.3
psycopg2-binary==2.9.6
sqlalchemy_utils==0.40.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
starlette-exporter==0.17.1




